+++ backend/database.py
@@ -15,5 +15,8 @@ Base = declarative_base()
 Base.query = db_session.query_property()
 
 def init_db():
-    import models
+    import models.data
+    import models.file
+    import models.user
+    
     Base.metadata.create_all(bind=engine)
