+++ backend/models/user.py
@@ -1,0 +1,9 @@
+from sqlalchemy import Column, Integer, String
+from database import Base
+
+
+class User(Base):
+    __tablename__ = "users"
+    id = Column(Integer, primary_key=True)
+    email = Column(String, nullable=False, index=True)
+    password = Column(String, nullable=False, index=True)
